# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Car Shop Documentation",
        "description": "Car Shop OpenApi description",
        "contact": {
            "email": "gaoun@globalistic.tech"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "car shop"
        }
    ],
    "paths": {
        "/car/get-car-by-id/{id}": {
            "get": {
                "tags": [
                    "Cars"
                ],
                "summary": "Get a specific car's information",
                "operationId": "getCarById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "status",
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "description": "data",
                                            "type": "object"
                                        },
                                        "message": {
                                            "description": "message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "status",
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "description": "data",
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "message": {
                                            "description": "message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "status",
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "description": "data",
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "message": {
                                            "description": "message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    }
                ]
            }
        },
        "/car/get-all-cars": {
            "get": {
                "tags": [
                    "Cars"
                ],
                "summary": "Get all cars information",
                "operationId": "getAllCars",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "name",
                        "required": false
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "sorting by newest to oldest => 1, oldest to newest =>2, A to Z => 3, Z to A =>4",
                        "required": false
                    },
                    {
                        "name": "minimum_price",
                        "in": "query",
                        "description": "filter by minimum price",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "max_price",
                        "in": "query",
                        "description": "filter by max price",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "brand_ids[]",
                        "in": "query",
                        "description": "filter by brand ids",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    },
                    {
                        "name": "type_ids[]",
                        "in": "query",
                        "description": "filter by type ids",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "number of data per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "status",
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "description": "data",
                                            "type": "object"
                                        },
                                        "message": {
                                            "description": "message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "status",
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "description": "data",
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "message": {
                                            "description": "message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "status",
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "description": "data",
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "message": {
                                            "description": "message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    }
                ]
            }
        },
...
        "/car/delete-car/{id}": {
            "delete": {
                "tags": [
                    "Cars"
                ],
                "summary": "Delete car",
                "operationId": "deleteCar",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "car_id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "status",
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "description": "data",
                                            "type": "object"
                                        },
                                        "message": {
                                            "description": "message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerToken": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}